rename CompanySocialImpactComparison
